requirements.txt
fastai==1.0.57
imageio==2.5.0
matplotlib==3.1.1
numpy==1.16.4
pandas==0.25.1
Pillow==6.1.0
scikit-image==0.15.0
scikit-learn==0.21.3
torch==1.2.0+cu92
torchvision==0.4.0+cu92
urllib3==1.25.3
